Add Path Tiling developer tool for deriving MultiBrush definitions
Adds a basic in-editor tool for creating segmented MultiBrush
definitions by using the Path Tiling Tool. This is primarily focused
at modders, but might form the basis of a user-generated MultiBrush
system in the future.

1. User places templates down on the map.
2. If needed, user defines the start, inner, and end types.
3. User traces over them using the Path Tiler Tool (perhaps
   individually or in a batch).
4. User clicks the button to dump the definitions to the OS clipboard.
5. User can then save these definitions to the tileset MiniYaml.

This is much faster and less error prone than trying to create the
definitions by hand.

By default, the developer tooling is only shown if the mod version is
"{DEV_VERSION}". A user can force the tooling to be shown or hidden
via a `Developer` field in world.yaml

Additional related changes:

- Add a checkbox to disable the tiling preview.
- Implement serialization of MultiBrushInfo back to MiniYaml.